components/mlops/api/model/requirements/requirement.txt
# ===============================================================================================================#
# Copyright 2024 Infosys Ltd.                                                                                    #
# Use of this source code is governed by Apache License Version 2.0 that can be found in the LICENSE file or at  #
# http://www.apache.org/licenses/                                                                                #
# ===============================================================================================================#

../lib/aicloudlibs-0.1.0-py3-none-any.whl
fastapi==0.83.0
pydantic==1.9.2
uvicorn
pymongo==3.13.0
python-dotenv==0.20.0
regex